+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Don't show the game link on the channel record page or for other players.
+
 ## [v21.3.0](https://github.com/lexicalunit/spellbot/releases/tag/v21.3.0) - 2026-06-18
 
 ### Added